requirements.txt
django==4.2.7
markdown==3.5.1
django-mdeditor==0.1.20